Entry,TextView: Deselect text on ::focus-out
We don't differentiate between foreground and backdropped selections,
which confuses/infuriates users, particularly in the FileChooser entry.
There's no design desire to introduce that differentiation in the theme.
The preferred fix is instead to clear the selection on focussing out.

This patch does that by changing the existing code that clears the
selection when the state changes and the widget is no longer :sensitive
to instead clear the selection if it's not the focus within its toplevel

https://gitlab.gnome.org/GNOME/gtk/issues/665#note_110313